feat: implement new frontend structure
- Added new layout and custom styles for the backend payload.
- Developed a comprehensive home page layout with sections for hero, services, projects, about, and contact.
- Introduced global styles and CSS variables for consistent theming across the application.
