+++ src/app/(frontend)/components/primitives/accordion/accordion.module.css
@@ -28,17 +28,13 @@
     display: flex;
     flex-direction: column;
     gap: var(--spacing-base);
-    height: 0;
+    max-height: 0;
     overflow: hidden;
-    transition: height 0.3s ease-in-out;
-    padding: 0;
-    margin: 0;
+    opacity: 0;
+    transition: all 1s ease-in-out;
 }
 
 .accordionBody.open {
-    height: auto;
-    padding: var(--spacing-base);
-    margin-top: var(--spacing-base);
-    margin-bottom: var(--spacing-base);
-    
+    max-height: 650px;
+    opacity: 1;
 }
