+++ src/app/(frontend)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.css";
-import Navigation from "@/components/navigation/Navigation";
+import "./styles/globals.css";
+import Navigation from "@/app/(frontend)/components/ui/navigation";
 
 const inter = Inter({ subsets: ["latin", "latin-ext"] });
