+++ src/app/(frontend)/styles/variables.css
